Fix a possibly wrong row offset on wrapped text
If the cursor was positioned in the left-most column and tmux wrapped
the shown text, tmux-easy-motion positioned the cursor on the first row
of the wrapped text (because the tmux `start-of-line` command jumps to
the start of wrapped text if the cursor is positioned in the first
column). This commit only executed `start-of-line` if the cursor is not
in the first column already.